+++ docs/changes.rst
@@ -40,6 +40,8 @@ Unreleased.
                      value=0.5 * img.quantum_range)
         img.save(filename='output.png')
 
+ - Added :meth:`Image.convex_hull() <wand.image.BaseImage.convex_hull>` method. Requires ImageMagick-7.0.10 or above.
+
 
 .. _changelog-0.6.3: